+++ src/QuickMarcEditor/QuickMarcEditorContainer.test.js
@@ -1,10 +1,12 @@
 import React from 'react';
-import { MemoryRouter } from 'react-router-dom';
+import { Router } from 'react-router-dom';
+import { createMemoryHistory } from 'history';
 import {
   render,
   cleanup,
   act,
   fireEvent,
+  screen,
 } from '@testing-library/react';
 import faker from 'faker';
 
@@ -33,7 +35,7 @@ const record = {
   fields: [],
 };
 
-const locations = {};
+const locations = [];
 
 const externalRecordPath = '/external/record/path';
 
@@ -43,8 +45,9 @@ const renderQuickMarcEditorContainer = ({
   action,
   wrapper,
   marcType = MARC_TYPES.BIB,
+  history = createMemoryHistory(),
 }) => (render(
-  <MemoryRouter>
+  <Router history={history}>
     <QuickMarcEditorContainer
       onClose={onClose}
       match={match}
@@ -54,7 +57,7 @@ const renderQuickMarcEditorContainer = ({
       marcType={marcType}
       externalRecordPath={externalRecordPath}
     />
-  </MemoryRouter>,
+  </Router>,
 ));
 
 describe('Given Quick Marc Editor Container', () => {
@@ -62,6 +65,7 @@ describe('Given Quick Marc Editor Container', () => {
   let instance;
 
   beforeEach(() => {
+    jest.clearAllMocks();
     instance = getInstance();
     mutator = {
       externalInstanceApi: {
@@ -93,6 +97,31 @@ describe('Given Quick Marc Editor Container', () => {
     });
 
     expect(mutator.quickMarcEditMarcRecord.GET).toHaveBeenCalled();
+  });
+
+  describe('when data cannot be fetched', () => {
+    const onClose = jest.fn();
+
+    beforeEach(async () => {
+      mutator.quickMarcEditMarcRecord.GET = jest.fn(() => Promise.reject());
+
+      await act(async () => {
+        renderQuickMarcEditorContainer({
+          mutator,
+          onClose,
+          action: QUICK_MARC_ACTIONS.DUPLICATE,
+          wrapper: QuickMarcEditWrapper,
+        });
+      });
+    });
+
+    it('should navigate back', () => {
+      expect(onClose).toHaveBeenCalled();
+    });
+
+    it('should not display Quick Marc Editor', () => {
+      expect(screen.queryByTestId('quick-marc-editor')).not.toBeInTheDocument();
+    });
   });
 
   it('should display Quick Marc Editor with fetched instance', async () => {