+++ src/QuickMarcEditor/QuickMarcEditorRows/QuickMarcEditorRows.js
@@ -1,5 +1,6 @@
 import React, {
   useCallback,
+  useRef,
 } from 'react';
 import PropTypes from 'prop-types';
 import {
@@ -7,8 +8,17 @@ import {
   useFormState,
 } from 'react-final-form';
 import { FieldArray } from 'react-final-form-arrays';
-import { useIntl } from 'react-intl';
+import {
+  FormattedMessage,
+  useIntl,
+} from 'react-intl';
+import isEqual from 'lodash/isEqual';
+import defer from 'lodash/defer';
 
+import {
+  Pluggable,
+  IfPermission,
+} from '@folio/stripes/core';
 import {
   TextField,
   IconButton,
@@ -21,6 +31,7 @@ import { MaterialCharsField } from './MaterialCharsField';
 import { PhysDescriptionField } from './PhysDescriptionField';
 import { FixedFieldFactory } from './FixedField';
 import { LocationField } from './LocationField';
+import { DeletedRowPlaceholder } from './DeletedRowPlaceholder';
 import {
   isReadOnly,
   hasIndicatorException,
@@ -44,40 +55,52 @@ const QuickMarcEditorRows = ({
   fields,
   type,
   subtype,
-  setDeletedRecords,
   mutators: {
     addRecord,
+    markRecordDeleted,
     deleteRecord,
     moveRecord,
+    restoreRecord,
   },
   marcType,
 }) => {
   const intl = useIntl();
   const { initialValues } = useFormState();
+  const containerRef = useRef(null);
+  const indexOfNewRow = useRef(null);
+  const newRowRef = useRef(null);
 
   const isNewRow = useCallback((row) => {
     return !initialValues.records.find(record => record.id === row.id);
   }, [initialValues.records]);
 
   const addNewRow = useCallback(({ target }) => {
-    addRecord({ index: parseInt(target.dataset.index, 10) });
+    const index = parseInt(target.dataset.index, 10);
+
+    indexOfNewRow.current = index + 1;
+    addRecord({ index });
+    defer(() => {
+      newRowRef.current.focus();
+    });
   }, [addRecord]);
 
   const deleteRow = useCallback(({ target }) => {
     const index = parseInt(target.dataset.index, 10);
+    const recordsLength = parseInt(target.dataset.recordsLength, 10);
+    const isLastRowDeleted = index === recordsLength - 1;
 
-    deleteRecord({ index });
-
-    if (!isNewRow(fields[index])) {
-      setDeletedRecords((prevDeletedRecords) => [
-        ...prevDeletedRecords,
-        {
-          index,
-          record: fields[index],
-        },
-      ]);
+    if (isNewRow(fields[index])) {
+      deleteRecord({ index });
+    } else {
+      markRecordDeleted({ index });
     }
-  }, [fields, deleteRecord, setDeletedRecords, isNewRow]);
+
+    if (!isLastRowDeleted) {
+      defer(() => {
+        containerRef.current.querySelector(`[name="records[${index + 1}].tag"]`).focus();
+      });
+    }
+  }, [fields, deleteRecord, markRecordDeleted, isNewRow, containerRef]);
 
   const moveRow = useCallback(({ target }) => {
     moveRecord({
@@ -86,18 +109,44 @@ const QuickMarcEditorRows = ({
     });
   }, [moveRecord]);
 
+  const restoreRow = useCallback((index) => {
+    restoreRecord({ index });
+  }, [restoreRecord]);
+
+  const processTagRef = useCallback(ref => {
+    if (!ref) return;
+    const index = parseInt(ref.dataset.index, 10);
+
+    if (indexOfNewRow.current === index) {
+      newRowRef.current = ref;
+    }
+  }, [indexOfNewRow, newRowRef]);
+
   return (
     <div
       id="quick-marc-editor-rows"
       data-testid="quick-marc-editor-rows"
+      ref={containerRef}
     >
-      <FieldArray name="records">
+      <FieldArray
+        name="records"
+        isEqual={isEqual}
+      >
         {({ fields: records }) => (
           records.map((name, idx) => {
             const recordRow = fields[idx];
 
             if (!recordRow) {
               return null;
+            }
+
+            if (recordRow._isDeleted) {
+              return (
+                <DeletedRowPlaceholder
+                  field={recordRow}
+                  restoreRow={() => restoreRow(idx)}
+                />
+              );
             }
 
             const isDisabled = isReadOnly(recordRow, action, marcType);
@@ -169,6 +218,8 @@ const QuickMarcEditorRows = ({
 
                 <div className={styles.quickMarcEditorRowTag}>
                   <Field
+                    inputRef={processTagRef}
+                    data-index={idx}
                     dirty={false}
                     ariaLabel={intl.formatMessage({ id: 'ui-quick-marc.record.field' })}
                     name={`${name}.tag`}
@@ -246,7 +297,6 @@ const QuickMarcEditorRows = ({
                     <LocationField
                       id={`location-field-${idx}`}
                       name={`${name}.content`}
-                      action={action}
                     />
                   )}
 
@@ -286,11 +336,19 @@ const QuickMarcEditorRows = ({
                         ariaLabel={intl.formatMessage({ id: 'ui-quick-marc.record.deleteField' })}
                         data-testid={`data-test-remove-row-${idx}`}
                         data-index={idx}
+                        data-records-length={records.length}
                         icon="trash"
                         onClick={deleteRow}
                       />
                     )
                   }
+                  <IfPermission perm="ui-quick-marc.quick-marc-authority-records.linkUnlink">
+                    <Pluggable
+                      type="find-authority"
+                    >
+                      <FormattedMessage id="ui-quick-marc.noPlugin" />
+                    </Pluggable>
+                  </IfPermission>
                 </div>
               </div>
             );
@@ -311,12 +369,14 @@ QuickMarcEditorRows.propTypes = {
     indicators: PropTypes.arrayOf(PropTypes.string),
     isProtected: PropTypes.bool.isRequired,
     content: PropTypes.oneOfType([PropTypes.string, PropTypes.object]).isRequired,
+    _isDeleted: PropTypes.bool,
   })),
-  setDeletedRecords: PropTypes.func.isRequired,
   mutators: PropTypes.shape({
     addRecord: PropTypes.func.isRequired,
     deleteRecord: PropTypes.func.isRequired,
+    markRecordDeleted: PropTypes.func.isRequired,
     moveRecord: PropTypes.func.isRequired,
+    restoreRecord: PropTypes.func.isRequired,
   }),
   marcType: PropTypes.oneOf(Object.values(MARC_TYPES)).isRequired,
 };