+++ src/QuickMarcEditor/QuickMarcEditorRows/QuickMarcEditorRows.test.js
@@ -7,6 +7,7 @@ import {
 } from '@testing-library/react';
 import { Form } from 'react-final-form';
 import arrayMutators from 'final-form-arrays';
+import defer from 'lodash/defer';
 
 import '@folio/stripes-acq-components/test/jest/__mock__';
 
@@ -15,7 +16,9 @@ import * as utils from './utils';
 import { QUICK_MARC_ACTIONS } from '../constants';
 import { MARC_TYPES } from '../../common/constants';
 
-const values = [
+jest.mock('lodash/defer', () => jest.fn());
+
+const initValues = [
   {
     id: '1',
     tag: '001',
@@ -57,10 +60,16 @@ const values = [
   },
 ];
 
-const addRecordMock = jest.fn();
+let values = [...initValues];
+const addRecordMock = jest.fn().mockImplementation(({ index }) => {
+  values.splice(index, 0, {
+    id: 'new-1',
+    content: '',
+  });
+});
 const deleteRecordMock = jest.fn();
 const moveRecordMock = jest.fn();
-const setDeletedRecordsMock = jest.fn();
+const markRecordDeletedMock = jest.fn();
 
 const renderQuickMarcEditorRows = (props = {}) => (render(
   <MemoryRouter>
@@ -68,7 +77,7 @@ const renderQuickMarcEditorRows = (props = {}) => (render(
       onSubmit={jest.fn()}
       mutators={arrayMutators}
       initialValues={{
-        records: values,
+        records: initValues,
       }}
       render={() => (
         <QuickMarcEditorRows
@@ -80,10 +89,10 @@ const renderQuickMarcEditorRows = (props = {}) => (render(
           mutators={{
             addRecord: addRecordMock,
             deleteRecord: deleteRecordMock,
+            markRecordDeleted: markRecordDeletedMock,
             moveRecord: moveRecordMock,
           }}
           subtype="test"
-          setDeletedRecords={setDeletedRecordsMock}
           {...props}
         />
       )}
@@ -92,6 +101,11 @@ const renderQuickMarcEditorRows = (props = {}) => (render(
 ));
 
 describe('Given QuickMarcEditorRows', () => {
+  beforeEach(() => {
+    values = [...initValues];
+    jest.clearAllMocks();
+  });
+
   afterEach(cleanup);
 
   it('should display row for each record value', () => {
@@ -142,8 +156,28 @@ describe('Given QuickMarcEditorRows', () => {
       expect(addRecordMock).toHaveBeenCalled();
     });
 
+    it('should focus the new `tag` field with a delay', async () => {
+      const { getAllByRole } = renderQuickMarcEditorRows();
+      const [addButton] = getAllByRole('button', { name: 'ui-quick-marc.record.addField' });
+
+      fireEvent.click(addButton);
+      expect(defer).toHaveBeenCalled();
+    });
+
+    describe('when the subfield is focused', () => {
+      it('should have a cursor at the end of the field value', () => {
+        const { getByTestId } = renderQuickMarcEditorRows();
+        const subfield = getByTestId('content-field-0');
+        const valueLength = subfield.value.length;
+        const spySetSelectionRange = jest.spyOn(subfield, 'setSelectionRange');
+
+        fireEvent.focus(subfield);
+        expect(spySetSelectionRange).toHaveBeenCalledWith(valueLength, valueLength);
+      });
+    });
+
     describe('and deleting a new row and saving', () => {
-      it('should mark the row as deleted', () => {
+      it('should not mark the row as deleted', () => {
         const {
           getAllByRole,
           getByTestId,
@@ -156,17 +190,18 @@ describe('Given QuickMarcEditorRows', () => {
 
         fireEvent.click(addButton);
         // delete button next to added row
-        const deleteButton = getAllByRole('button', { name: 'ui-quick-marc.record.addField' })[1];
+        const deleteButton = getAllByRole('button', { name: 'ui-quick-marc.record.deleteField' })[0];
 
         fireEvent.click(deleteButton);
 
-        expect(setDeletedRecordsMock).not.toHaveBeenCalled();
+        expect(markRecordDeletedMock).not.toHaveBeenCalled();
+        expect(deleteRecordMock).toHaveBeenCalled();
       });
     });
   });
 
   describe('when deleting rows', () => {
-    it('should call setDeletedRecords 2 times', () => {
+    it('should call markRecordDeleted 2 times', () => {
       const { getAllByTestId } = renderQuickMarcEditorRows();
 
       const testIdx1 = 1;
@@ -177,17 +212,15 @@ describe('Given QuickMarcEditorRows', () => {
       fireEvent.click(deleteIcon1[1]);
       fireEvent.click(deleteIcon2[1]);
 
-      expect(setDeletedRecordsMock).toHaveBeenCalledTimes(2);
+      expect(markRecordDeletedMock).toHaveBeenCalledTimes(2);
     });
 
-    it('should handle deleteRecord', () => {
+    it('should focus the `tag` field that is located after the deleted field with a delay', async () => {
       const { getAllByRole } = renderQuickMarcEditorRows();
-
       const [deleteButton] = getAllByRole('button', { name: 'ui-quick-marc.record.deleteField' });
 
       fireEvent.click(deleteButton);
-
-      expect(deleteRecordMock).toHaveBeenCalled();
+      expect(defer).toHaveBeenCalled();
     });
   });