+++ src/QuickMarcEditor/utils.js
@@ -93,7 +93,6 @@ const getCreateMarcRecordDefaultFields = (instanceRecord) => {
     }
 
     if (tag === '852') {
-      field.content = '$a';
       field.indicators = ['\\', '\\'];
     }
 
@@ -191,6 +190,16 @@ export const formatMarcRecordByQuickMarcAction = (marcRecord, action) => {
   }
 
   return marcRecord;
+};
+
+export const addInternalFieldProperties = (marcRecord) => {
+  return {
+    ...marcRecord,
+    records: marcRecord.records.map(record => ({
+      ...record,
+      _isDeleted: false,
+    })),
+  };
 };
 
 export const hydrateMarcRecord = marcRecord => ({
@@ -297,7 +306,7 @@ export const validateLeader = (prevLeader = '', leader = '', marcType = MARC_TYP
 };
 
 export const getLocationValue = (value) => {
-  const matches = value?.match(/\$b\s([^$\s]+\/?)+/) || [];
+  const matches = value?.match(/\$b\s+([^$\s]+\/?)+/) || [];
 
   return matches[0] || '';
 };
@@ -471,6 +480,17 @@ export const deleteRecordByIndex = (index, state) => {
   return records;
 };
 
+export const markDeletedRecordByIndex = (index, state) => {
+  const records = [...state.formState.values.records];
+
+  records[index] = {
+    ...records[index],
+    _isDeleted: true,
+  };
+
+  return records;
+};
+
 export const reorderRecords = (index, indexToSwitch, state) => {
   const records = [...state.formState.values.records];
 
@@ -479,12 +499,24 @@ export const reorderRecords = (index, indexToSwitch, state) => {
   return records;
 };
 
-export const restoreRecordAtIndex = (index, record, state) => {
+export const restoreRecordAtIndex = (index, state) => {
   const records = [...state.formState.values.records];
 
-  records.splice(index, 0, record);
+  records[index] = {
+    ...records[index],
+    _isDeleted: false,
+  };
 
   return records;
+};
+
+export const removeDeletedRecords = (formValues) => {
+  const { records } = formValues;
+
+  return {
+    ...formValues,
+    records: records.filter(record => !record._isDeleted),
+  };
 };
 
 export const removeFieldsForDuplicate = (formValues) => {